\name{gbggm-package}
\alias{gbggm-package}
\alias{gbggm}
\docType{package}
\title{
\packageTitle{gbggm}
}
\description{
\packageDescription{gbggm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gbggm}
\packageIndices{gbggm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gbggm}

Maintainer: \packageMaintainer{gbggm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
